lookie!
so, i'm just using codemirror's included compression script for the js
(which just calls uglify). normally that uses their webservice, but
because we're doing so much, we're using a local uglify. it would be
_really_ sloppy to try to squeeze into package.json, so for right now,
cd into public/codemirror_libs/ and run the minify-targets.sh (but
read it first, just in case!). the output codemirror.bundle.js should
then just live in public root.